requirements.txt
# LRM-MIA
# Python Requirements File

# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
tqdm>=4.65.0

# Deep Learning Frameworks
torch>=2.0.0
transformers>=4.30.0

# Sentence Embeddings & NLP
sentence-transformers>=2.2.0
nltk>=3.8.0
tokenizers>=0.21.0

# String Similarity & Edit Distance
python-Levenshtein>=0.21.0

# Machine Learning & Evaluation
scikit-learn>=1.3.0

# Excel File Support
openpyxl>=3.1.0
xlrd>=2.0.1

# API Client for LLM Services
openai>=1.0.0
google-genai>=1.57.0
httpx>=0.28.0

# Tokenizers
tiktoken>=0.11.0

# Data Processing
datasets>=4.0.0
pyarrow>=22.0.0

# Evaluation Metrics
rouge-score>=0.1.2

# Configuration & Utilities
pydantic>=2.0.0
PyYAML>=6.0.0
regex>=2024.0.0

# Optional: GPU Monitoring
# nvitop>=1.5.0
# nvidia-ml-py>=13.580.0

# Optional: Visualization
# matplotlib>=3.8.0

# Optional: For better performance
# accelerate>=0.20.0
# bitsandbytes>=0.39.0
